Moves from vector_math to vector_math_64

+++ examples/raw/spinning_square.dart
@@ -27,7 +27,7 @@ void beginFrame(double timeStamp) {
   // composite
   final double devicePixelRatio = sky.view.devicePixelRatio;
   sky.Rect sceneBounds = new sky.Rect.fromLTWH(0.0, 0.0, sky.view.width * devicePixelRatio, sky.view.height * devicePixelRatio);
-  Float32List deviceTransform = new Float32List(16)
+  Float64List deviceTransform = new Float64List(16)
     ..[0] = devicePixelRatio
     ..[5] = devicePixelRatio
     ..[10] = 1.0
